create model file
