+++ internal/service/searchindex/model_search_index.go
@@ -1,0 +1,144 @@
+package searchindex
+
+import (
+	"context"
+	"encoding/json"
+	"log"
+
+	"github.com/go-test/deep"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/retry"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/mongodb/terraform-provider-mongodbatlas/internal/common/conversion"
+	"go.mongodb.org/atlas-sdk/v20240530002/admin"
+)
+
+func flattenSearchIndexSynonyms(synonyms []admin.SearchSynonymMappingDefinition) []map[string]any {
+	synonymsMap := make([]map[string]any, len(synonyms))
+	for i, s := range synonyms {
+		synonymsMap[i] = map[string]any{
+			"name":              s.Name,
+			"analyzer":          s.Analyzer,
+			"source_collection": s.Source.Collection,
+		}
+	}
+	return synonymsMap
+}
+
+func expandSearchIndexSynonyms(d *schema.ResourceData) []admin.SearchSynonymMappingDefinition {
+	var synonymsList []admin.SearchSynonymMappingDefinition
+	if vSynonyms, ok := d.GetOk("synonyms"); ok {
+		for _, s := range vSynonyms.(*schema.Set).List() {
+			synonym := s.(map[string]any)
+			synonymsDoc := admin.SearchSynonymMappingDefinition{
+				Name:     synonym["name"].(string),
+				Analyzer: synonym["analyzer"].(string),
+				Source: admin.SynonymSource{
+					Collection: synonym["source_collection"].(string),
+				},
+			}
+			synonymsList = append(synonymsList, synonymsDoc)
+		}
+	}
+	return synonymsList
+}
+
+func marshalSearchIndex(fields any) (string, error) {
+	bytes, err := json.Marshal(fields)
+	return string(bytes), err
+}
+
+func unmarshalSearchIndexMappingFields(str string) (map[string]any, diag.Diagnostics) {
+	fields := map[string]any{}
+	if str == "" {
+		return fields, nil
+	}
+	if err := json.Unmarshal([]byte(str), &fields); err != nil {
+		return nil, diag.Errorf("cannot unmarshal search index attribute `mappings_fields` because it has an incorrect format")
+	}
+	return fields, nil
+}
+
+func unmarshalSearchIndexFields(str string) ([]map[string]any, diag.Diagnostics) {
+	fields := []map[string]any{}
+	if str == "" {
+		return fields, nil
+	}
+	if err := json.Unmarshal([]byte(str), &fields); err != nil {
+		return nil, diag.Errorf("cannot unmarshal search index attribute `fields` because it has an incorrect format")
+	}
+
+	return fields, nil
+}
+
+func unmarshalSearchIndexAnalyzersFields(str string) ([]admin.AtlasSearchAnalyzer, diag.Diagnostics) {
+	fields := []admin.AtlasSearchAnalyzer{}
+	if str == "" {
+		return fields, nil
+	}
+	if err := json.Unmarshal([]byte(str), &fields); err != nil {
+		return nil, diag.Errorf("cannot unmarshal search index attribute `analyzers` because it has an incorrect format")
+	}
+	return fields, nil
+}
+
+func validateSearchIndexMappingDiff(k, old, newStr string, d *schema.ResourceData) bool {
+	var j, j2 any
+
+	if old == "" {
+		old = "{}"
+	}
+
+	if newStr == "" {
+		newStr = "{}"
+	}
+
+	if err := json.Unmarshal([]byte(old), &j); err != nil {
+		log.Printf("[ERROR] cannot unmarshal old search index mapping json %v", err)
+	}
+	if err := json.Unmarshal([]byte(newStr), &j2); err != nil {
+		log.Printf("[ERROR] cannot unmarshal new search index mapping json %v", err)
+	}
+	if diff := deep.Equal(&j, &j2); diff != nil {
+		log.Printf("[DEBUG] deep equal not passed: %v", diff)
+		return false
+	}
+
+	return true
+}
+
+func validateSearchAnalyzersDiff(k, old, newStr string, d *schema.ResourceData) bool {
+	var j, j2 any
+
+	if old == "" {
+		old = "{}"
+	}
+
+	if newStr == "" {
+		newStr = "{}"
+	}
+
+	if err := json.Unmarshal([]byte(old), &j); err != nil {
+		log.Printf("[ERROR] cannot unmarshal old search index analyzer json %v", err)
+	}
+	if err := json.Unmarshal([]byte(newStr), &j2); err != nil {
+		log.Printf("[ERROR] cannot unmarshal new search index analyzer json %v", err)
+	}
+	if diff := deep.Equal(&j, &j2); diff != nil {
+		log.Printf("[DEBUG] deep equal not passed: %v", diff)
+		return false
+	}
+
+	return true
+}
+
+func resourceSearchIndexRefreshFunc(ctx context.Context, clusterName, projectID, indexID string, connV2 *admin.APIClient) retry.StateRefreshFunc {
+	return func() (any, string, error) {
+		searchIndex, _, err := connV2.AtlasSearchApi.GetAtlasSearchIndex(ctx, projectID, clusterName, indexID).Execute()
+		if err != nil {
+			return nil, "ERROR", err
+		}
+		status := conversion.SafeString(searchIndex.Status)
+		return searchIndex, status, nil
+	}
+}
